refactor: Context API 사용한 부분들 제거
