+++ src/main.tsx
@@ -2,12 +2,9 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
 import './index.css';
-import CalculatorProvider from '@/store/CalculatorProvider';
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
 	<React.StrictMode>
-		<CalculatorProvider>
-			<App/>
-		</CalculatorProvider>
+		<App/>
 	</React.StrictMode>,
 );
